+++ hobj/benchmarks/mut_highvar_benchmark.py
@@ -3,7 +3,7 @@ from typing import List, Dict
 
 from hobj.benchmarks.binary_classification.benchmark import LearningCurveBenchmark, LearningCurveBenchmarkConfig, TargetSubtaskData
 from hobj.benchmarks.binary_classification.simulation import BinaryClassificationSubtask, BinaryClassificationSubtaskResult
-from hobj.data.behavior import load_highvar_behavior
+from hobj.loaders.behavior import load_highvar_behavior
 from hobj.images import MutatorHighVarImageset
