+++ hobj/benchmarks/mut_oneshot_benchmark.py
@@ -5,7 +5,7 @@ from typing import Dict, List
 from hobj.benchmarks.generalization.benchmark import GeneralizationBenchmarkConfig, GeneralizationBenchmark, GeneralizationSessionResult
 from hobj.benchmarks.generalization.estimator import GeneralizationStatistics
 from hobj.benchmarks.generalization.simulator import GeneralizationSubtask
-from hobj.data.behavior import load_oneshot_behavior
+from hobj.loaders.behavior import load_oneshot_behavior
 from hobj.images import MutatorOneShotImageset
 from mref import ImageRef
