+++ tests/test_load_behavior.py
@@ -1,4 +1,4 @@
-from hobj.data.behavior import load_highvar_behavior, load_oneshot_behavior
+from hobj.loaders.behavior import load_highvar_behavior, load_oneshot_behavior
 
 
 def test_load_highvar():
